+++ R/tribe.R
@@ -9,6 +9,13 @@
 #' @param obj
 #' An object.
 #' 
+#' @param x
+#' A list (of attributes) to be \code{untribe}d. 
+#' 
+#' @param keep_obj
+#' logical. If \code{TRUE}, \code{obj} is passed as an attribute to 
+#' the result (useful in combination of \code{untribe}). 
+#' 
 #' @param value
 #' An appropriate named list of attributes, or \code{NULL}. 
 #' 
@@ -23,12 +30,26 @@
 #' @importFrom bazar nlist
 #' @export
 #' 
+#' @examples 
+#' library(lplyr)
+#' A <- c(x = 1, y = 2, z = 3) %>% 
+#'   define(package = "trib?")
+#' A %>% 
+#'   tribe(keep_obj = TRUE) %@>% 
+#'   mutate(package = "tribe") %>% 
+#'   untribe()
+#' 
 tribe <- 
-function(obj)
+function(obj, 
+         keep_obj = FALSE)
 {
   at <- attributes(obj)
   if (is.null(at)) {
-    return(bazar::nlist())
+    at <- bazar::nlist()
+  }
+  if (keep_obj) {
+    #attributes(obj) <- NULL
+    attr(at, ".obj_tribe") <- obj
   }
   at
 }
@@ -42,5 +63,18 @@ function(obj)
 function(obj, value)
 {
   attributes(obj) <- if (bazar::is.empty(value)) NULL else value
+  obj
+}
+
+
+#' @export
+#' @rdname tribe
+#' 
+untribe <- 
+function(x)
+{
+  obj <- attr(x, ".obj_tribe")
+  attr(x, ".obj_tribe") <- NULL
+  attributes(obj) <- x
   obj
 }