+++ NEWS.md
@@ -1,7 +1,23 @@
+# tribe 0.1.5
+
+* New argument 'keep_obj' in tribe()
+* New function untribe()
+
+
+# tribe 0.1.4
+
+* Updated authors in the DESCRIPTION file. 
+
+
+# tribe 0.1.3
+
+* stick_to() and unstick() added. 
+
+
 # tribe 0.1.2
 
 * The pipe '%at>%' is renamed '%@>%'. 
-* Dplyr verbs for list manipulations and 'mutate_which' function 
+* Dplyr verbs for list manipulations and mutate_which() function 
 are transferred to the 'lplyr' package
 * Infix attribute accessor "%@%" is added (borrowed from package 'purrr')